initial implementation of variant API access declaratively.
added the ability to have `androidComponents` in the toml declarative
file. This applies only to be able to set properties, none of the APIs
related to artifacts will ever be available.

So far, only intrinsic types and Property<*> are supported, support
for other types like MapProperty<>, ListProperty<> will come in a
subsequent CL.

Added simple tests for `application` and `library` androidComponents
extensions.

Test: unit tests.
Bug: N/A