Reworked variant selection.
Following Alex Golubev suggestion in a previous commit, reworked
the variant selection to be simpler by using :
   [androidComponents.onVariants.all] to target all variants.
   [androidComponents.onVariants.XXX] to target variants maching
   the XXX name.

It is not possible to implement regular expression matching even
using toml dotted keys as a number of characters like ?, [, ]
cannot be used and makes building a regular expression in a key
not possible. We might revisit later using a different formalism.

Tests: Added unit tests.
Bug: N/A