Add ability to include other toml buildscripts into a build script
A Toml buildscript can define an array of relative path to other Toml
buildscripts which will be recursively parsed depth first and applied to
extension objects. The paths are relative to the current buildscript.
Processing a Toml buildscript creates a side effect in form of cached
extension objects. All build script parsing now shares the same cache.
The scripts are processed in the order they are declared in the array
and the nested file is processed before the file that includes it.
Plugins are applied first to make the extension objects available to
configure before the nested file processing ensue. The result of this
order of processing is that the declarations in the included file are
overriden by the file that includes it. Similarly, the files that
precedes in the incude array have it's declarations overriden by the
files that are at the higer index in the include array in case of
conflict. Current test include loading a single library project with the
nested buildscripts without verification. More tests will be added as we
build our test infrastructure for verifying whether the DSL is applied
correctly or not.