Support for calling methods with multiple parameters.
Added support for calling methods that have more than one paramters,
format is :

missingDimensionStrategy = [ "contentType", "demo" ]

which will be translated into setMissingDimensionStrategy("contentType", "demo")

Also added support for container support. When dealing with an assignment like

signingConfig = "debug"

and we realize that signingConfig is not a String but a type T, we lookup if there
is a property that is a NamedDomainObjectContainer<T> in the context. If we find one,
we loop up by name using the assignment value (debug in this case) and assing the
return value.

Test: converted the NowInAndroid project.